Added Prompt for Delete
Upon clicking the delete button a prompt shows up asking if the user is
certain that they wish to delete the configuration.

Also added in random functions that do not work. I was trying to check
for overwriting files that were already saved. This did not work - I ran
into a problem where it would immediately go into the compare function
before even getting the name from the user.