Save functionality restored.
There was a problem with saving a file and then loading right after
would crash the app. The problem was that the Scroll View was attempting
to add mLinearLayout again, thus throwing an IllegalStateException. The
Fix was to simply move the scroll view to the OnCreate function so that
it is only called once.